Add files via upload
As the cruise comes to an end, you’ll disembark with lasting memories of an unforgettable experience. From the stunning views of Mumbai’s landmarks and coastline to the luxurious comfort of the yacht, the City Cruise Tour offers a unique way to see the city from a perspective few get to experience.

Whether you’re a lover of luxury or simply looking for a different way to discover Mumbai, this cruise tour promises to be an experience like no other.